scripts/requirements-transcript-translate.txt
py-googletrans==4.0.0
httpx==0.27.2
